refactor: Refactored the helper functions in order to work dynamically
